refactor!: change how west dependencies are loaded
Refactor the image so the west dependencies are not contained in the image, but loaded using west init and west update. This makes the image more reusable to different projects with different dependencies.

BREAKING CHANGE: The folder structure have changed to /workdir/project and west needs to be initialized and updated before building a project the first time.